Refactor: Group classes into packages
