+++ src/main/java/bobbybot/exceptions/TooManyArgumentsException.java
@@ -1,4 +1,4 @@
-package bobbybot;
+package bobbybot.exceptions;
 
 /**
  * Exception for when too many arguments are given to a command
