+++ src/main/java/bobbybot/util/Ui.java
@@ -1,4 +1,4 @@
-package bobbybot;
+package bobbybot.util;
 
 /**
  * Encapsulates all user-facing interactions
