+++ src/test/java/TaskListTest.java
@@ -1,4 +1,4 @@
-import bobbybot.TaskList;
+import bobbybot.util.TaskList;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
